+++ attributes/default.rb
@@ -1,7 +1,7 @@
 default['yum-epel']['repos'] =
   value_for_platform(
     %w(almalinux redhat centos oracle rocky) => {
-      '>= 8.0' => epel_8_repos,
+      '>= 8.0' => epel_repos,
       '~> 7.0' =>
         %w(
           epel